+++ Types/Emoji.Sequences/get_All.ps1
@@ -1,0 +1,12 @@
+<#
+.SYNOPSIS
+    Gets Emoji Sequences
+.DESCRIPTION
+    Gets all loaded Emoji Sequences
+#>
+param()
+foreach ($property in $this.psobject.properties) {
+    if (-not $property.IsInstance) { continue }
+    if ($property -isnot [psnoteproperty]) { continue }
+    $property.Value
+}