+++ Types/Emoji/get_Demo.ps1
@@ -1,0 +1,13 @@
+<#
+.SYNOPSIS
+    Gets module demos
+.DESCRIPTION
+    Gets demo files related to a module.
+.EXAMPLE
+    $Posh.Demos
+.LINK
+    https://github.com/StartAutomating/ShowDemo
+#>
+$this |
+    Split-Path |
+        Get-ChildItem -Recurse -Filter *.demo.ps1