+++ Types/Emoji/get_Sequences.ps1
@@ -1,0 +1,13 @@
+<#
+.SYNOPSIS
+    Emoji Sequences
+.DESCRIPTION
+    Gets the Emoji Sequences Collection
+#>
+if (-not $this.'.Sequences') {
+    $this  | Add-Member NoteProperty '.Sequences' (
+        [PSCustomObject]@{PSTypeName='Emoji.Sequences'}
+    ) -Force
+}
+
+$this.'.Sequences'