+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## Emoji 0.1.3:
+
+* Improved Formatting
+  * Now with Color (and better padding) (#46)
+  * Adding Custom formats (#47)
+  * Adding PowerShell formats (#49)
+* Emoji Sequences:
+  * Get-Emoji -Sequence (#55)
+  * Emoji.Sequences.All (#41)
+  * Emoji.Sequences.Add (#39)
+  * Emoji.Sequences.Remove (#40)
+  * Emoji.get/set Sequence (#35)
+  * Sequence Formatting (#54)
+* Added `Emoji.get_Demos` (#42)
+* Added `/Demos/Emoji.demo.ps1` (#43)
+* Set-Emoji
+  * Added -ScriptBlock (#51)
+  * Defaulting Value (#52)
+
+---
+
 ## Emoji 0.1.2:
 
 ### Updating Emoji: